main.py: added remove_gaps function.
The function aims to remove gaps of more than (1 * fps) frames, between
an element in the list of frames of a repetition and the previous one.
If so, the list will be split and only the largest one kept. The same is
repeated for all the reps list of frames.

This way a repetition has no gap and can be identified as a (start_sec, end_sec) tuple,
for the next step of classification.